renamed Debug.js to Debugged.js
